Improved Grades Zipping & Extraction
Removed the call to reduce and instead replaced the previous map calls with reduce, concat and map. The result is shorter.

I'd like to use 'Zip' or 'ZipWith', but from my research into Ramda.JS and RxJS, it doesn't seem like the right tool for the job.

The next thing I'd like to improve is the part of this code where I call sort twice and then reduce. I believe that could be shortened.